それぞれのstorageにアクセスするための関数をモジュールにしてファイル分け
